backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.3.0
Flask-MySQLdb==1.0.1